extract update state into hooks
