+++ apps/desktop/src/routes/__root.tsx
@@ -13,6 +13,7 @@ import { useStorageListener } from "@desktop/hooks/use-app-storage";
 import { useDeeplinkListener } from "@desktop/hooks/use-deeplink-listener";
 import { useShortcutListener } from "@desktop/hooks/use-shortcut-listener";
 import { useTheme } from "@desktop/hooks/use-theme";
+import { useUpdateListener } from "@desktop/hooks/use-update-listener";
 import { i18n } from "@desktop/i18n";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { createRootRoute, Outlet } from "@tanstack/react-router";
@@ -34,6 +35,7 @@ function RootComponent() {
   useShortcutListener();
   useDeeplinkListener();
   useStorageListener();
+  useUpdateListener();
 
   return (
     <I18nextProvider i18n={i18n}>
